For Loop Guantlet
The For-Loop Gauntlet

The goal of this assignment is to make you a master of utilizing for
loops. Complete all the for loop challenges on the paper. Please read
each challenge carefully and make sure that your for loops display
exactly what is asked. Do not move on to the next challenge until an
instructor has verified that your previous challenge is correct. If you
finish them all, try the bonus challenges.

Single For-Loops
Write a for loop to do each the following:

1. Display all numbers from 0 to 100
2. Display all numbers from 100 to 0
3. Display all even numbers from 2 to 100
4. Display all odd numbers from 1 to 99
5. Display all numbers from 1 to 500. If the number is odd, print “odd”
next to the number. If the number is even, print “even” next to the
number.
	Ex:	1 odd
		2 even
		3 odd
		4 even
		5 odd...etc.
6. Display all multiples of 7 from 0 to 777.
7. Print all the years you were alive and how old you were in each.
e.g. “In 1979, i was 2 years old.” (for a really old person)

Nested For-Loops (a for loop inside of a for loop)
Write a nested for loop to do the following:

1. Display this output:
	0 0
	0 1
	0 2
	1 0
	1 1
	1 2
	2 0
	2 1
	2 2

2. Display the numbers 1 through 9 in a 3x3 square grid like this:
	1 2 3
	4 5 6
	7 8 9
3. Display the numbers 1 through 100 in a 10x10 square grid
4. Display the following output:
	*
	**
	***
	****
	*****
	******